Fix owner issue
